Product mapping - minor refactoring.
