+++ src/module-vsbridge-indexer-catalog/Index/Mapping/Attribute.php
@@ -2,39 +2,28 @@
 
 namespace Divante\VsbridgeIndexerCatalog\Index\Mapping;
 
+use Divante\VsbridgeIndexerCatalog\Index\Mapping\Attribute\SwatchMapping;
 use Divante\VsbridgeIndexerCore\Api\MappingInterface;
 use Divante\VsbridgeIndexerCore\Api\Mapping\FieldInterface;
-use Divante\VsbridgeIndexerCore\Index\Mapping\GeneralMapping;
-use Magento\Framework\Event\ManagerInterface as EventManager;
 
 /**
  * Class Attribute
  */
 class Attribute implements MappingInterface
 {
-
     /**
-     * @var EventManager
+     * @var SwatchMapping
      */
-    private $eventManager;
-
-    /**
-     * @var GeneralMapping
-     */
-    private $generalMapping;
+    private $swatchMapping;
 
     /**
      * Attribute constructor.
      *
-     * @param GeneralMapping $generalMapping
-     * @param EventManager $eventManager
+     * @param SwatchMapping $generalMapping
      */
-    public function __construct(
-        GeneralMapping $generalMapping,
-        EventManager $eventManager
-    ) {
-        $this->eventManager = $eventManager;
-        $this->generalMapping = $generalMapping;
+    public function __construct(SwatchMapping $generalMapping)
+    {
+        $this->swatchMapping = $generalMapping;
     }
 
     /**
@@ -125,19 +114,10 @@ class Attribute implements MappingInterface
                 'value' => ['type' => FieldInterface::TYPE_TEXT],
                 'label' => ['type' => FieldInterface::TYPE_TEXT],
                 'sort_order' => ['type' => FieldInterface::TYPE_INTEGER],
-                'swatch' => $this->generalMapping->getSwatchProperties(),
+                'swatch' => $this->swatchMapping->get(),
             ]
         ];
 
-        $mapping = ['properties' => $properties];
-        $mappingObject = new \Magento\Framework\DataObject();
-        $mappingObject->setData($mapping);
-
-        $this->eventManager->dispatch(
-            'elasticsearch_attribute_mapping_properties',
-            ['mapping' => $mappingObject]
-        );
-
-        return $mappingObject->getData();
+        return ['properties' => $properties];
     }
 }
