+++ src/module-vsbridge-indexer-catalog/Index/Mapping/Product.php
@@ -1,13 +1,17 @@
 <?php
+/**
+ * @package  Divante\VsbridgeIndexerCatalog
+ * @author Agata Firlejczyk <afirlejczyk@divante.pl>
+ * @copyright 2019 Divante Sp. z o.o.
+ * @license See LICENSE_DIVANTE.txt for license details.
+ */
 
 namespace Divante\VsbridgeIndexerCatalog\Index\Mapping;
 
 use Divante\VsbridgeIndexerCore\Api\Mapping\FieldInterface;
 use Divante\VsbridgeIndexerCore\Api\MappingInterface;
 use Divante\VsbridgeIndexerCore\Index\Mapping\GeneralMapping;
-use Divante\VsbridgeIndexerCatalog\Model\Attributes\ConfigurableAttributes;
 use Divante\VsbridgeIndexerCatalog\Model\ResourceModel\Product\LoadAttributes;
-use Magento\Framework\Event\ManagerInterface as EventManager;
 
 /**
  * Class Product
@@ -15,19 +19,14 @@ use Magento\Framework\Event\ManagerInterface as EventManager;
 class Product extends AbstractMapping implements MappingInterface
 {
     /**
-     * @var EventManager
-     */
-    private $eventManager;
-
-    /**
      * @var GeneralMapping
      */
     private $generalMapping;
 
     /**
-     * @var array
+     * @var StockMapping
      */
-    private $properties;
+    private $stockMapping;
 
     /**
      * @var LoadAttributes
@@ -35,30 +34,36 @@ class Product extends AbstractMapping implements MappingInterface
     private $resourceModel;
 
     /**
-     * @var ConfigurableAttributes
+     * @var array
      */
-    private $configurableAttributes;
+    private $properties;
+
+    /**
+     * @var FieldMappingInterface[]
+     */
+    private $additionalMapping = [];
 
     /**
      * Product constructor.
      *
-     * @param EventManager $eventManager
      * @param GeneralMapping $generalMapping
-     * @param ConfigurableAttributes $configurableAttributes
+     * @param StockMapping $stockMapping
      * @param LoadAttributes $resourceModel
      * @param array $staticFieldMapping
+     * @param array $additionalMapping
      */
     public function __construct(
-        EventManager $eventManager,
         GeneralMapping $generalMapping,
-        ConfigurableAttributes $configurableAttributes,
+        StockMapping $stockMapping,
         LoadAttributes $resourceModel,
-        array $staticFieldMapping
+        array $staticFieldMapping,
+        array $additionalMapping
     ) {
-        $this->eventManager = $eventManager;
+        $this->stockMapping = $stockMapping;
         $this->generalMapping = $generalMapping;
         $this->resourceModel = $resourceModel;
-        $this->configurableAttributes = $configurableAttributes;
+        $this->additionalMapping = $additionalMapping;
+
         parent::__construct($staticFieldMapping);
     }
 
@@ -70,25 +75,16 @@ class Product extends AbstractMapping implements MappingInterface
         if (null === $this->properties) {
             $allAttributesMapping = $this->getAllAttributesMappingProperties();
             $commonMappingProperties = $this->getCommonMappingProperties();
-
-            $childrenMapping = $this->getChildrenAttributeMappings($allAttributesMapping);
-            $childrenMapping = array_merge($childrenMapping, $commonMappingProperties);
-
             $attributesMapping = array_merge($allAttributesMapping, $commonMappingProperties);
 
             $properties = $this->getCustomProperties();
-            $properties['configurable_children'] = ['properties' => $childrenMapping];
+            $properties['configurable_children'] = ['properties' => $attributesMapping];
             $properties = array_merge($properties, $attributesMapping);
             $properties = array_merge($properties, $this->generalMapping->getCommonProperties());
 
             $mapping = ['properties' => $properties];
             $mappingObject = new \Magento\Framework\DataObject();
             $mappingObject->setData($mapping);
-
-            $this->eventManager->dispatch(
-                'elasticsearch_product_mapping_properties',
-                ['mapping' => $mappingObject]
-            );
 
             $this->properties = $mappingObject->getData();
         }
@@ -116,30 +112,12 @@ class Product extends AbstractMapping implements MappingInterface
     }
 
     /**
-     * @param array $allAttributes
-     *
      * @return array
      */
-    private function getChildrenAttributeMappings(array $allAttributes = [])
-    {
-        $list = [];
-
-        foreach ($this->configurableAttributes->getChildrenRequiredAttributes() as $field) {
-            if (isset($allAttributes[$field])) {
-                $list[$field] = $allAttributes[$field];
-            }
-        }
-
-        return $list;
-    }
-
-    /**
-     * @return array
-     */
-    private function getCommonMappingProperties()
+    private function getCommonMappingProperties(): array
     {
         $attributesMapping = [];
-        $attributesMapping['stock']['properties'] = $this->generalMapping->getStockMapping();
+        $attributesMapping['stock']['properties'] = $this->stockMapping->get();
         $attributesMapping['media_gallery'] = [
             'properties' => [
                 'type' => ['type' => FieldInterface::TYPE_TEXT],
@@ -169,156 +147,15 @@ class Product extends AbstractMapping implements MappingInterface
      */
     private function getCustomProperties(): array
     {
-        return [
-            'attribute_set_id' => ['type' => FieldInterface::TYPE_LONG],
-            'bundle_options' => $this->getBundleOptionsMapping(),
-            'product_links' => $this->getProductLinksMapping(),
-            'configurable_options' => $this->getConfigurableOptionsMapping(),
-            'category' => $this->getCategoryMapping(),
-            'custom_options' => $this->getCustomOptionsMapping(),
-            'tier_prices' => $this->getTierPricesMapping(),
-        ];
-    }
+        $customProperties = ['attribute_set_id' => ['type' => FieldInterface::TYPE_LONG]];
 
-    /**
-     * @return array
-     */
-    private function getProductLinksMapping(): array
-    {
-        return [
-            'properties' => [
-                'linked_product_type' => ['type' => FieldInterface::TYPE_TEXT],
-                'linked_product_sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                'sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                'position' => ['type' => FieldInterface::TYPE_LONG],
-            ],
-        ];
-    }
+        foreach ($this->additionalMapping as $propertyName => $properties) {
+            if ($properties instanceof FieldMappingInterface) {
+                $customProperties[$propertyName] = $properties->get();
+            }
+        }
 
-    /**
-     * @return array
-     */
-    private function getConfigurableOptionsMapping(): array
-    {
-        return [
-            'properties' => [
-                'label' => ['type' => FieldInterface::TYPE_TEXT],
-                'id' => ['type' => FieldInterface::TYPE_LONG],
-                'product_id' => ['type' => FieldInterface::TYPE_LONG],
-                'attribute_code' => ['type' => FieldInterface::TYPE_TEXT],
-                'attribute_id' => ['type' => FieldInterface::TYPE_LONG],
-                'position' => ['type' => FieldInterface::TYPE_LONG],
-                'values' => [
-                    'properties' => [
-                        'value_index' => ['type' => FieldInterface::TYPE_KEYWORD],
-                        'label' => ['type' => FieldInterface::TYPE_TEXT],
-                        'swatch' => $this->generalMapping->getSwatchProperties(),
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @return array
-     */
-    private function getCategoryMapping(): array
-    {
-        return  [
-            'type' => 'nested',
-            'properties' => [
-                'category_id' => ['type' => FieldInterface::TYPE_LONG],
-                'position' => ['type' => FieldInterface::TYPE_LONG],
-                'name' => [
-                    'type' => FieldInterface::TYPE_TEXT,
-                    'fields' => [
-                        'keyword' => [
-                            'type' => FieldInterface::TYPE_KEYWORD,
-                            'ignore_above' => 256,
-                        ]
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @return array
-     */
-    private function getBundleOptionsMapping(): array
-    {
-        return [
-            'properties' => [
-                'option_id' => ['type' => FieldInterface::TYPE_LONG],
-                'position' => ['type' => FieldInterface::TYPE_LONG],
-                'title' => ['type' => FieldInterface::TYPE_TEXT],
-                'sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                'product_links' => [
-                    'properties' => [
-                        'id' => ['type' => FieldInterface::TYPE_LONG],
-                        'is_default' => ['type' => FieldInterface::TYPE_BOOLEAN],
-                        'qty' => ['type' => FieldInterface::TYPE_DOUBLE],
-                        'can_change_quantity' => ['type' => FieldInterface::TYPE_BOOLEAN],
-                        'price' => ['type' => FieldInterface::TYPE_DOUBLE],
-                        'price_type' => ['type' => FieldInterface::TYPE_TEXT],
-                        'position' => ['type' => FieldInterface::TYPE_LONG],
-                        'sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                    ],
-                ],
-            ]
-        ];
-    }
-
-    /**
-     * @return array
-     */
-    private function getCustomOptionsMapping(): array
-    {
-        return [
-            'properties' => [
-                'image_size_x' => ['type' => FieldInterface::TYPE_TEXT],
-                'image_size_y' => ['type' => FieldInterface::TYPE_TEXT],
-                'file_extension' => ['type' => FieldInterface::TYPE_TEXT],
-                'is_require' => ['type' => FieldInterface::TYPE_BOOLEAN],
-                'max_characters' => ['type' => FieldInterface::TYPE_TEXT],
-                'option_id' => ['type' => FieldInterface::TYPE_LONG],
-                'price' => ['type' => FieldInterface::TYPE_DOUBLE],
-                'price_type' => ['type' => FieldInterface::TYPE_TEXT],
-                'sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                'sort_order' => ['type' => FieldInterface::TYPE_INTEGER],
-                'title' => ['type' => FieldInterface::TYPE_TEXT],
-                'type' => ['type' => FieldInterface::TYPE_TEXT],
-                'values' => [
-                    'properties' => [
-                        'sku' => ['type' => FieldInterface::TYPE_KEYWORD],
-                        'price' => ['type' => FieldInterface::TYPE_DOUBLE],
-                        'title' => ['type' => FieldInterface::TYPE_TEXT],
-                        'price_type' => ['type' => FieldInterface::TYPE_TEXT],
-                        'sort_order' => ['type' => FieldInterface::TYPE_INTEGER],
-                        'option_type_id' => ['type' => FieldInterface::TYPE_INTEGER],
-                    ]
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * @return array
-     */
-    private function getTierPricesMapping(): array
-    {
-        return [
-            'properties' => [
-                'customer_group_d' => ['type' => FieldInterface::TYPE_INTEGER],
-                'qty' => ['type' => FieldInterface::TYPE_DOUBLE],
-                'value' => ['type' => FieldInterface::TYPE_DOUBLE],
-                'extension_attributes' => [
-                    'properties' => [
-                        'website_id' => ['type' => FieldInterface::TYPE_SHORT]
-                    ],
-                ],
-            ],
-        ];
+        return $customProperties;
     }
 
     /**
