+++ src/module-vsbridge-indexer-catalog/Model/Category/ApplyCategorySlug.php
@@ -9,7 +9,7 @@
 namespace Divante\VsbridgeIndexerCatalog\Model\Category;
 
 use Divante\VsbridgeIndexerCatalog\Api\ApplyCategorySlugInterface;
-use Divante\VsbridgeIndexerCatalog\Api\Data\CatalogConfigurationInterface;
+use Divante\VsbridgeIndexerCatalog\Api\CatalogConfigurationInterface;
 use Divante\VsbridgeIndexerCatalog\Api\SlugGeneratorInterface;
 
 /**
